filepath function issue
